Added struct b_offs to solid_base. Refactored solid_base, solid_draw, and mapc to accomodate b_offs. Temporarily commented out the glMultiTexture conditional. Fixed long type warning in sync.c under OSX. Fixed Ogg Vorbis warning static callbacks. Added timer feature to mapc. Added CSV feature to mapc. Removed "visible geom" output from mapc output, as offs is more relevant. Removed uniq_geoms tweak from mapc. Added uniq_mtrl tweak to mapc (just for symmetry, not for optimization.)
git-svn-id: https://s.snth.net/svn/neverball/branches/gles@3527 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ share/sync.c
@@ -30,7 +30,7 @@ void sync_init(void)
 
 void sync_init(void)
 {
-    long swap = 1;
+    GLint swap = 1;
     CGLSetParameter(CGLGetCurrentContext(),  kCGLCPSwapInterval, &swap);
 }
